+++ odpi-ext/dpi-ext.h
@@ -1,0 +1,7 @@
+#ifndef ODPI_EXT_H
+#define ODPI_EXT_H
+#include "dpi.h"
+
+int dpi_ext_dpiStmt_getFnCode(dpiStmt *stmt, uint16_t *sqlfncode);
+
+#endif